[TASK] Starts with the implementation of a new random generator concept

+++ LightBulb/Function/XorShfGenerator.hpp
@@ -44,7 +44,7 @@ public:
 
 	double next()
 	{
-		return (float)(*this)() / max();
+		return (double)(*this)() / max();
 	}
 };
